requirements.txt
asgiref==3.8.1
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
folium==0.19.5
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
pip==21.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==58.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
xyzservices==2025.1.0
